refactor(api): move uncancellable error
into src/team/domain/errors.js
