+++ api/src/team/domain/errors.js
@@ -1,0 +1,10 @@
+import { DomainError } from '../../shared/domain/errors.js';
+
+export class UncancellableCertificationCenterInvitationError extends DomainError {
+  constructor(
+    message = "L'invitation à ce centre de certification ne peut pas être annulée.",
+    code = 'UNCANCELLABLE_CERTIFICATION_CENTER_INVITATION_CODE',
+  ) {
+    super(message, code);
+  }
+}
